// src/components/Context/CartContext.tsx
"use client"
import { CartResponse } from "@/interfaces/Cart";
import { createContext, ReactNode, useEffect, useState} from "react";

export const CartContext = createContext<{
 cartData: CartResponse|null,
 setCartData: (value:CartResponse|null)=> void,
 isLoading: boolean,
 setIsLoading: (value:boolean)=>void,
 getCart:()=> void,
}>({
 cartData:null,
 setCartData:()=>{},
 isLoading:false,
setIsLoading:()=>{}, 
getCart() {},
});
export default function CartContextProvider({children}:{children:ReactNode}){
  const [cartData,setCartData]=useState<CartResponse|null>(null);
  const [isLoading,setIsLoading]=useState<boolean>(true);
//   const [userId,setUserId]=useState<string>('');

   async function getCart(){
    const response = await fetch ('https://ecommerce.routemisr.com/api/v1/cart',{
        method:"GET",
      headers:{
  token: localStorage.getItem("userToken") || ""
}

    });
    const data :CartResponse = await response.json();
    setCartData(data);
    // if(cartData?.data.cartOwner){
    //  localStorage.setItem('userId',cartData?.data.cartOwner)
    // }
     if(data?.data?.cartOwner){
    localStorage.setItem('userId', data.data.cartOwner)
  }
    ;
    setIsLoading(false);
    console.log(data);
    
   }
   useEffect(()=>{
    getCart()
   },[])


    return <>
       <CartContext.Provider value={{cartData,setCartData,isLoading,setIsLoading,getCart}}>
        {children}
    </CartContext.Provider>
    </>
  
}